requirements.txt
django
django-taggit
django-widget-tweaks